+++ src/elements/BarCodeElement.js
@@ -129,4 +129,14 @@ export default class BarCodeElement extends DocElement {
         this.addCommandForChangedParameterName(parameter, newParameterName, 'content', cmdGroup);
         this.addCommandForChangedParameterName(parameter, newParameterName, 'printIf', cmdGroup);
     }
+
+    /**
+     * Returns class name.
+     * This can be useful for introspection when the class names are mangled
+     * due to the webpack uglification process.
+     * @returns {string}
+     */
+    getClassName() {
+        return 'BarCodeElement';
+    }
 }