+++ src/elements/TableElement.js
@@ -590,6 +590,16 @@ export default class TableElement extends DocElement {
     static getColumnMinWidth() {
         return 20;
     }
+
+    /**
+     * Returns class name.
+     * This can be useful for introspection when the class names are mangled
+     * due to the webpack uglification process.
+     * @returns {string}
+     */
+    getClassName() {
+        return 'TableElement';
+    }
 }
 
 TableElement.border = {